requirements.txt
langchain
langchain-community 
langgraph 
fastapi
python-dotenv
langchain-groq
pydantic
streamlit 
pandas
numpy
uvicorn
langsmith


-e .